move get_windows from gen to multistream
