+++ nnsvs/bin/generate.py
@@ -10,9 +10,8 @@ import torch
 from hydra.utils import to_absolute_path
 from nnmnkwii.datasets import FileSourceDataset
 from nnsvs.base import PredictionType
-from nnsvs.gen import get_windows
 from nnsvs.logger import getLogger
-from nnsvs.multistream import multi_stream_mlpg
+from nnsvs.multistream import get_windows, multi_stream_mlpg
 from nnsvs.train_util import NpyFileSource
 from omegaconf import DictConfig, OmegaConf
 from tqdm import tqdm
